requirements.txt
aiogram==2.22.2
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pillow==10.3.0
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
